+++ db/migrate/20160718145122_remove_settings_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveSettingsFromUsers < ActiveRecord::Migration
+  def change
+    remove_column :users, :settings, :text
+  end
+end