+++ db/migrate/20160718152051_remove_status_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveStatusFromUsers < ActiveRecord::Migration
+  def change
+    remove_column :users, :status, :string, null: false, default: 'pending'
+  end
+end